Create leathanachNuaver10.java
**most up to date - working to return a word based on frequency counts using algorithm for comparison**

Android java file that takes in characters into an edittext box and based on these characters will return the predicted word based on what is populated in the arrays. Here input "aba", "abh", "bea", "bai" will return the predicted word.

**Algorithm being used**
A number of characters as entered into the input field (should work for any number) are compared against substrings of each word of a non-fada array of Irish words. Substring length will be the amount of charaters entered. On a match occurring the accurate corresponding fada word in the fada array (based on index position) will be stored in a results array as a possible output. All resulting possible output values will be compared against their frequency count in the count array (a results position count variable will be needed to track the index position for comparison) and the highest frequency value/word will be returned. Each time a higher frequency value is found is it stored in the highestprob variable.